+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Unreleased
 ### Added
 - The `warn-octal-off` flag.
+- The `info` command now accepts one or more argument for specific aspects of the program state.
+- The `info` command can now print information about the program's call stack.
 
 ### Changed
 - The `--no-ret-warn` flag has been renamed to `--warn-return-off`.